requirements.txt
Django==1.4.1
SQLAlchemy==0.8.1
South==0.7.6
beautifulsoup4==4.1.3
boto==2.9.2
django-analytical==0.15.0
django-extensions==1.1.1
django-mysqlpool==0.1-8
django-registration==0.8
django-registration-email==0.5.1
django-storages==1.1.8
html5lib==0.95
nltk==2.0.4
numpy==1.7.1
python-memcached==1.51
pytz==2013b
PyYAML==3.10
re2==0.2.20
simplejson==3.1.2
six==1.3.0
stripe==1.7.9